refactor(frontend): extract YieldChartSegment into its own module
Move the memoized YieldChartSegment sub-component (and its
YieldSegmentPayload type) out of YieldOverview into
src/pages/YieldChartSegment.tsx, further shrinking the page file. React
event types are imported directly instead of via the React namespace. The
now-unused Tooltip, ContextMenuIndicator, contextMenuIndicatorHostSx, and
memo imports are dropped from the page. No behavior change.
